+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/ui/settings/SettingsConfigurable.kt
@@ -2,7 +2,8 @@ package com.github.blarc.gitlab.template.lint.plugin.ui.settings
 
 import com.github.blarc.gitlab.template.lint.plugin.GitlabLintBundle.message
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
-import com.github.blarc.gitlab.template.lint.plugin.ui.SettingsForm
+import com.github.blarc.gitlab.template.lint.plugin.settings.ProjectSettings
+import com.intellij.openapi.components.service
 import com.intellij.openapi.options.Configurable
 import com.intellij.openapi.project.Project
 import org.jetbrains.annotations.Nls
@@ -32,28 +33,41 @@ class SettingsConfigurable(val project: Project) : Configurable {
 
     override fun isModified(): Boolean {
         val settings = AppSettings.instance
+        val projectSettings = project.service<ProjectSettings>()
         val gitlabUrl = settingsForm?.gitlabUrlTF
+
         if (settings != null && gitlabUrl != null) {
-            return !settingsForm!!.gitlabTokenTF.contentEquals(settings.getGitlabToken(gitlabUrl)) || settingsForm!!.gitlabRemotesTable.isModified(settings)
+            return !settingsForm!!.gitlabTokenTF.contentEquals(settings.getGitlabToken(gitlabUrl)) ||
+                    settingsForm!!.gitlabRemotesTable.isModified(settings) ||
+                    !settingsForm!!.gitlabUrlTF.contentEquals(projectSettings.gitlabUrl) ||
+                    !settingsForm!!.gitlabRemoteTF.contentEquals(projectSettings.remote)
         }
         return false
     }
 
     override fun apply() {
         val settings = AppSettings.instance
+        val projectSettings = project.service<ProjectSettings>()
         val gitlabUrl = settingsForm?.gitlabUrlTF
+
         if (settings != null && gitlabUrl != null) {
             settings.saveGitlabToken(settingsForm!!.gitlabTokenTF!!, gitlabUrl)
             settingsForm!!.gitlabRemotesTable.commit(settings)
+            projectSettings.gitlabUrl = settingsForm!!.gitlabUrlTF!!
+            projectSettings.remote = settingsForm!!.gitlabRemoteTF!!
         }
     }
 
     override fun reset() {
         val settings = AppSettings.instance
+        val projectSettings = project.service<ProjectSettings>()
         val gitlabUrl = settingsForm?.gitlabUrlTF
+
         if (settings != null && gitlabUrl != null) {
             settingsForm!!.gitlabTokenTF = settings.getGitlabToken(gitlabUrl)
             settingsForm!!.gitlabRemotesTable.tableModel.remotesList = settings.remotesMap.toList().toMutableList()
+            settingsForm!!.gitlabUrlTF = projectSettings.gitlabUrl
+            settingsForm!!.gitlabRemoteTF = projectSettings.remote
         }
     }